Extract code_mode runtime into dedicated crate
