+++ codex-rs/code-mode/src/code_mode_bridge.js
