+++ codex-rs/core/Cargo.toml
@@ -32,6 +32,7 @@ codex-app-server-protocol = { workspace = true }
 codex-apply-patch = { workspace = true }
 codex-async-utils = { workspace = true }
 codex-client = { workspace = true }
+codex-code-mode = { workspace = true }
 codex-config = { workspace = true }
 codex-shell-command = { workspace = true }
 codex-skills = { workspace = true }
@@ -81,7 +82,6 @@ rmcp = { workspace = true, default-features = false, features = [
     "schemars",
     "server",
 ] }
-rusty_v8 = { workspace = true }
 schemars = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
